refactor: use CreateIndexesOperation for createIndex
