+++ lib/collection.js
@@ -3,6 +3,7 @@
 const deprecate = require('util').deprecate;
 const deprecateOptions = require('./utils').deprecateOptions;
 const checkCollectionName = require('./utils').checkCollectionName;
+const prepareCreateIndex = require('./utils').prepareCreateIndex;
 const ObjectID = require('./core').BSON.ObjectID;
 const MongoError = require('./core').MongoError;
 const toError = require('./utils').toError;
@@ -35,7 +36,6 @@ const updateDocuments = require('./operations/common_functions').updateDocuments
 const AggregateOperation = require('./operations/aggregate');
 const BulkWriteOperation = require('./operations/bulk_write');
 const CountDocumentsOperation = require('./operations/count_documents');
-const CreateIndexOperation = require('./operations/create_index');
 const CreateIndexesOperation = require('./operations/create_indexes');
 const DeleteManyOperation = require('./operations/delete_many');
 const DeleteOneOperation = require('./operations/delete_one');
@@ -1240,14 +1240,16 @@ Collection.prototype.createIndex = function(fieldOrSpec, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
   options = options || {};
 
-  const createIndexOperation = new CreateIndexOperation(
-    this.s.db,
+  const prepared = prepareCreateIndex(fieldOrSpec, options);
+  const createIndexesOperation = new CreateIndexesOperation(
+    this,
     this.collectionName,
-    fieldOrSpec,
-    options
+    prepared.indexSpec,
+    prepared.options,
+    true
   );
 
-  return executeOperation(this.s.topology, createIndexOperation, callback);
+  return executeOperation(this.s.topology, createIndexesOperation, callback);
 };
 
 /**
@@ -1296,7 +1298,12 @@ Collection.prototype.createIndexes = function(indexSpecs, options, callback) {
 
   if (typeof options.maxTimeMS !== 'number') delete options.maxTimeMS;
 
-  const createIndexesOperation = new CreateIndexesOperation(this, indexSpecs, options);
+  const createIndexesOperation = new CreateIndexesOperation(
+    this,
+    this.collectionName,
+    indexSpecs,
+    options
+  );
 
   return executeOperation(this.s.topology, createIndexesOperation, callback);
 };
