+++ lib/db.js
@@ -3,6 +3,7 @@
 const EventEmitter = require('events').EventEmitter;
 const inherits = require('util').inherits;
 const getSingleProperty = require('./utils').getSingleProperty;
+const prepareCreateIndex = require('./utils').prepareCreateIndex;
 const CommandCursor = require('./command_cursor');
 const handleCallback = require('./utils').handleCallback;
 const filterOptions = require('./utils').filterOptions;
@@ -36,7 +37,7 @@ const AddUserOperation = require('./operations/add_user');
 const CollectionsOperation = require('./operations/collections');
 const CommandOperation = require('./operations/command');
 const CreateCollectionOperation = require('./operations/create_collection');
-const CreateIndexOperation = require('./operations/create_index');
+const CreateIndexesOperation = require('./operations/create_indexes');
 const DropCollectionOperation = require('./operations/drop').DropCollectionOperation;
 const DropDatabaseOperation = require('./operations/drop').DropDatabaseOperation;
 const ExecuteDbAdminCommandOperation = require('./operations/execute_db_admin_command');
@@ -748,9 +749,16 @@ Db.prototype.createIndex = function(name, fieldOrSpec, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
   options = options ? Object.assign({}, options) : {};
 
-  const createIndexOperation = new CreateIndexOperation(this, name, fieldOrSpec, options);
+  const prepared = prepareCreateIndex(fieldOrSpec, options);
+  const createIndexesOperation = new CreateIndexesOperation(
+    this,
+    name,
+    prepared.indexSpec,
+    prepared.options,
+    true
+  );
 
-  return executeOperation(this.s.topology, createIndexOperation, callback);
+  return executeOperation(this.s.topology, createIndexesOperation, callback);
 };
 
 /**
