+++ lib/operations/create_indexes.js
@@ -5,35 +5,23 @@ const defineAspects = require('./operation').defineAspects;
 const CommandOperationV2 = require('./command_v2');
 const MongoError = require('../core').MongoError;
 const maxWireVersion = require('../core/utils').maxWireVersion;
-const ReadPreference = require('../core').ReadPreference;
 
-/**
- * Creates one or more indexes on the collection.
- *
- * @class
- */
 class CreateIndexesOperation extends CommandOperationV2 {
   /**
-   * Construct a CreateIndexes operation.
-   *
-   * @param {Collection} collection parent collection
-   * @param {string|object} indexSpecs Defines the indexes.
-   * @param {object} [options] Optional settings. See Collection.prototype.createIndexes for a list of options.
+   * @ignore
    */
-  constructor(collection, indexSpecs, options) {
-    super(collection, options);
-    this.collection = collection;
+  constructor(parent, name, indexSpecs, options, singular) {
+    super(parent, options);
+    this.name = name;
     this.indexSpecs = indexSpecs;
+    this.singular = singular;
   }
 
   /**
-   * Execute the operation.
-   *
-   * @param {any} server
-   * @param {Collection~resultCallback} [callback] The command result callback
+   * @ignore
    */
   execute(server, callback) {
-    const options = Object.assign({}, this.options, { readPreference: ReadPreference.PRIMARY });
+    const options = this.options;
     const indexSpecs = this.indexSpecs;
 
     const serverWireVersion = maxWireVersion(server);
@@ -63,7 +51,7 @@ class CreateIndexesOperation extends CommandOperationV2 {
     }
 
     const cmd = {
-      createIndexes: this.collection.collectionName,
+      createIndexes: this.name,
       indexes: indexSpecs
     };
 
@@ -80,8 +68,14 @@ class CreateIndexesOperation extends CommandOperationV2 {
       cmd.commitQuorum = options.commitQuorum;
     }
 
-    this.decorateWithCollation = false;
-    super.executeCommand(server, cmd, callback);
+    super.executeCommand(server, cmd, (err, result) => {
+      if (err) {
+        callback(err);
+        return;
+      }
+
+      callback(null, this.singular ? indexSpecs[0].name : result);
+    });
   }
 }
