+++ lib/utils.js
@@ -187,6 +187,41 @@ var normalizeHintField = function normalizeHintField(hint) {
   return finalHint;
 };
 
+const keysToOmit = new Set([
+  'name',
+  'key',
+  'commitQuorum',
+  'writeConcern',
+  'w',
+  'wtimeout',
+  'j',
+  'fsync',
+  'readPreference',
+  'session'
+]);
+
+const prepareCreateIndex = function(fieldOrSpec, options) {
+  const indexParameters = parseIndexOptions(fieldOrSpec);
+  // Generate the index name
+  const name = typeof options.name === 'string' ? options.name : indexParameters.name;
+  // Set up the index
+  const indexes = { name, key: indexParameters.fieldHash };
+  // merge all the options
+  for (let optionName in options) {
+    if (!keysToOmit.has(optionName)) {
+      indexes[optionName] = options[optionName];
+    }
+  }
+  return {
+    indexSpec: [indexes],
+    options: {
+      commitQuorum: options.commitQuorum,
+      readPreference: options.readPreference,
+      session: options.session
+    }
+  };
+};
+
 /**
  * Create index name based on field spec
  *
@@ -744,6 +779,7 @@ module.exports = {
   checkCollectionName,
   toError,
   formattedOrderClause,
+  prepareCreateIndex,
   parseIndexOptions,
   normalizeHintField,
   handleCallback,
