+++ test/functional/index.test.js
@@ -1364,6 +1364,7 @@ describe('Indexes', function() {
             const db = client.db('test');
             const collection = db.collection('commitQuorum');
             testCommand(db, collection, (err, result) => {
+              expect(err).to.exist;
               expect(err.message).to.equal(
                 '`commitQuorum` option for `createIndexes` not supported on servers < 4.4'
               );
